requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
mysqlclient==2.1.1
pymongo==4.3.3
pytz==2022.7.1
sqlparse==0.2.4
tzdata==2022.7
